vprdb/vpr_systems/netvlad/i_dataset.py
#  Copyright (c) 2023, Ivan Moskalenko, Anastasiia Kornilova
#
#  Licensed under the Apache License, Version 2.0 (the "License");
#  you may not use this file except in compliance with the License.
#  You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.
from pathlib import Path
from PIL import Image
from torch.utils.data import Dataset

from vprdb.vpr_systems.utils import input_transform


class IDataset(Dataset):
    def __init__(self, images: list[Path], resize=(480, 640)):
        self.images = images
        self.transform = input_transform(resize)

    def __len__(self):
        return len(self.images)

    def __getitem__(self, index):
        img = Image.open(self.images[index])
        img = self.transform(img)

        return img, index
