requirements.txt
faiss_cpu==1.7.3
joblib==1.2.0
mrob==0.0.11
networkx==3.0
nptyping==2.5.0
numpy==1.24.2
open3d==0.17.0
opencv_python==4.7.0.72
Pillow==9.4.0
pytorch_lightning==2.0.0
scikit_learn==1.2.2
scipy==1.10.1
tensorboardX==2.6
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
